backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.0
elevenlabs==0.2.1
firebase-admin==6.4.0
requests==2.31.0
pydantic==2.5.0
aiofiles==23.2.0
gunicorn==21.2.0
slowapi==0.1.9
